changed file name class name and constructor name
